requirements.txt
click
scikit_learn
diffusers==0.26.3
einops==0.8.1
matplotlib==3.8.3
matplotlib==3.7.2
numpy==1.24.1
packaging==25.0
Pillow==9.3.0
Pillow==11.2.1
piqa==1.3.2
scipy==1.15.3
seaborn==0.13.2
simple_parsing==0.1.5
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
